Implement clone_shards and finish_node_install functions for database management and installation completion messages
